# web/auth/login.go
package auth

import (
	"net/http"
	"time"

	"github.com/SkyPanel/SkyPanel/v3"
	"github.com/SkyPanel/SkyPanel/v3/middleware"
	"github.com/SkyPanel/SkyPanel/v3/models"
	"github.com/SkyPanel/SkyPanel/v3/response"
	"github.com/SkyPanel/SkyPanel/v3/scopes"
	"github.com/SkyPanel/SkyPanel/v3/services"
	"github.com/gin-contrib/sessions"
	"github.com/gin-gonic/gin"
)

// @Summary User Login
// @Description Authenticates a user and starts a session
// @Accept json
// @Produce json
// @Param request body LoginRequestData true "Login Request"
// @Success 200 {object} LoginResponse
// @Failure 400 {object} SkyPanel.ErrorResponse
// @Tags Auth
// @Router /auth/login [post]
func LoginPost(c *gin.Context) {
	db := middleware.GetDatabase(c)
	us := &services.User{DB: db}

	request := &LoginRequestData{}

	err := c.BindJSON(request)
	if response.HandleError(c, err, http.StatusBadRequest) {
		return
	}

	user, otpNeeded, err := us.ValidateLogin(request.Email, request.Password)
	if response.HandleError(c, err, http.StatusBadRequest) {
		return
	}

	if otpNeeded {
		userSession := sessions.Default(c)
		userSession.Set("user", user.Email)
		userSession.Set("time", time.Now().Unix())
		err = userSession.Save()
		if response.HandleError(c, err, http.StatusInternalServerError) {
			return
		}
		c.JSON(http.StatusOK, &LoginResponse{
			OtpNeeded: true,
		})
		return
	}

	createSession(c, user)
}

// @Summary OTP Verification
// @Description Verifies an OTP code for a login attempt
// @Accept json
// @Produce json
// @Param request body OtpRequestData true "OTP Request"
// @Success 200 {object} LoginResponse
// @Failure 400 {object} SkyPanel.ErrorResponse
// @Tags Auth
// @Router /auth/otp [post]
func OtpPost(c *gin.Context) {
	db := middleware.GetDatabase(c)
	us := &services.User{DB: db}

	request := &OtpRequestData{}

	err := c.BindJSON(request)
	if response.HandleError(c, err, http.StatusBadRequest) {
		return
	}

	userSession := sessions.Default(c)
	email := userSession.Get("user").(string)
	timestamp := userSession.Get("time").(int64)

	if email == "" {
		response.HandleError(c, SkyPanel.ErrInvalidSession, http.StatusBadRequest)
		return
	}

	if timestamp < time.Now().Unix()-300 {
		userSession.Clear()
		_ = userSession.Save()
		response.HandleError(c, SkyPanel.ErrSessionExpired, http.StatusBadRequest)
		return
	}

	user, err := us.ValidOtp(email, request.Token)
	if response.HandleError(c, err, http.StatusBadRequest) {
		return
	}

	createSession(c, user)
}

func createSession(c *gin.Context, user *models.User) {
	db := middleware.GetDatabase(c)
	ps := &services.Permission{DB: db}
	ss := &services.Session{DB: db}

	perms, err := ps.GetForUserAndServer(user.ID, "")
	if response.HandleError(c, err, http.StatusInternalServerError) {
		return
	}

	// Aggregate scopes from individual permissions and roles
	allScopes := perms.Scopes

	if user.RoleId != nil {
		rs := &services.Role{DB: db}
		role, err := rs.Get(*user.RoleId)
		if err == nil && role != nil {
			for _, s := range role.Scopes {
				allScopes = scopes.AddScope(allScopes, scopes.GetScope(s))
			}
		}
	}

	if !scopes.ContainsScope(allScopes, scopes.ScopeLogin) {
		response.HandleError(c, SkyPanel.ErrLoginNotPermitted, http.StatusForbidden)
		return
	}

	session, err := ss.CreateForUser(user)
	if response.HandleError(c, err, http.StatusInternalServerError) {
		return
	}

	data := &LoginResponse{}
	data.Scopes = allScopes

	secure := false
	if c.Request.TLS != nil {
		secure = true
	}

	maxAge := int(time.Hour / time.Second)

	c.SetCookie("puffer_auth", session, maxAge, "/", "", secure, true)
	c.SetCookie("puffer_auth_expires", "", maxAge, "/", "", secure, false)

	c.JSON(http.StatusOK, data)
}

type LoginRequestData struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}

type LoginResponse struct {
	Scopes    []*scopes.Scope `json:"scopes,omitempty"`
	OtpNeeded bool            `json:"otpNeeded,omitempty"`
}

type OtpRequestData struct {
	Token string `json:"token"`
}