# web/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "Authenticates a user and starts a session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User Login",
                "parameters": [
                    {
                        "description": "Login Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginRequestData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/otp": {
            "post": {
                "description": "Verifies an OTP code for a login attempt",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "OTP Verification",
                "parameters": [
                    {
                        "description": "OTP Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.OtpRequestData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/register": {
            "post": {
                "description": "Creates a new user account if registration is enabled",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Registration Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.registerRequestData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/daemon": {
            "get": {
                "security": [
                    {
                        "OAuth2Application": [
                            "none"
                        ]
                    }
                ],
                "description": "Check to see if the daemon is online or not",
                "tags": [
                    "Daemon Root"
                ],
                "summary": "Check daemon status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DaemonRunning"
                        }
                    }
                }
            },
            "head": {
                "security": [
                    {
                        "OAuth2Application": [
                            "none"
                        ]
                    }
                ],
                "description": "Check to see if the daemon is online or not",
                "tags": [
                    "Daemon Root"
                ],
                "summary": "Check daemon status",
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/daemon/features": {
            "get": {
                "security": [
                    {
                        "OAuth2Application": [
                            "none"
                        ]
                    }
                ],
                "description": "Gets the features that the node supports, like it's OS and environments",
                "tags": [
                    "Daemon Root"
                ],
                "summary": "Get features of the node",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Features"
                        }
                    }
                }
            }
        },
...
        "/daemon/system": {
            "get": {
                "security": [
                    {
                        "OAuth2Application": [
                            "none"
                        ]
                    }
                ],
                "description": "Gets detailed system information including CPU, memory, and disk",
                "tags": [
                    "Daemon Root"
                ],
                "summary": "Get system information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SystemInfo"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "DaemonRunning": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "DiskInfo": {
            "type": "object",
            "properties": {
                "free": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                },
                "used": {
                    "type": "integer"
                },
                "usedPercent": {
                    "type": "number"
                }
            }
        },
...
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/Error"
                }
            }
        },
...
        "Features": {
            "type": "object",
            "properties": {
                "arch": {
                    "type": "string"
                },
                "environments": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "features": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "os": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "Scopes": {
            "type": "object",
            "properties": {
                "forServer": {
                    "type": "boolean"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "SystemInfo": {
            "type": "object",
            "properties": {
                "arch": {
                    "type": "string"
                },
                "cpuCores": {
                    "type": "integer"
                },
                "cpuModel": {
                    "type": "string"
                },
                "cpuThreads": {
                    "type": "integer"
                },
                "cpuUsage": {
                    "type": "number"
                },
                "disks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/DiskInfo"
                    }
                },
                "hostname": {
                    "type": "string"
                },
                "memoryFree": {
                    "type": "integer"
                },
                "memoryTotal": {
                    "type": "integer"
                },
                "memoryUsed": {
                    "type": "integer"
                },
                "networkBytesRecv": {
                    "type": "integer"
                },
                "networkBytesSent": {
                    "type": "integer"
                },
                "os": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "platformVersion": {
                    "type": "string"
                },
                "uptime": {
                    "type": "integer"
                }
            }
        },
...
        "auth.LoginRequestData": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "auth.LoginResponse": {
            "type": "object",
            "properties": {
                "otpNeeded": {
                    "type": "boolean"
                },
                "scopes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Scopes"
                    }
                }
            }
        },
        "auth.OtpRequestData": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "auth.registerRequestData": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 5
                }
            }
        },
...
    }
}